requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
betterproto==2.0.0b6
boto3==1.37.28
botocore==1.37.28
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-storages==1.14.6
dnspython==2.7.0
email_validator==2.2.0
grpclib==0.4.7
gunicorn==23.0.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
id==1.5.0
idna==3.10
in-toto-attestation==0.9.3
jmespath==1.0.1
kagglehub==0.3.7
markdown-it-py==3.0.0
mdurl==0.1.2
model-signing==0.2.0
multidict==6.1.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
rfc3161-client==0.1.2
rfc8785==0.1.4
rich==13.9.4
s3transfer==0.11.4
securesystemslib==1.2.0
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tuf==5.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.9.0
